Fix some matching file regressions vs master

+++ extern/dolphin/include/dolphin/axfx.h
@@ -153,9 +153,9 @@ void AXFXReverbStdCallback(struct AXFX_BUFFERUPDATE* bufferUpdate,
                            struct AXFX_REVERBSTD* reverb);
 
 // axfx.c
-void* AXFXAllocFunction(unsigned long long size);
+void* AXFXAllocFunction(unsigned long size);
 void AXFXFreeFunction(void* ptr);
-void AXFXSetHooks(void* (*alloc_hook)(unsigned long long),
+void AXFXSetHooks(void* (*alloc_hook)(unsigned long),
                   void (*free_hook)(void*));
 
 #endif // _DOLPHIN_AXFX_H_
